requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
